Extract tool_surface tests into tool_surface/tests.rs
